Move loadConfiguration round-trip tests into cli.MainConfigurationLoadingTest
The earlier commit (63ef722) put the new tests in
net.ladenthin.bitcoinaddressfinder.MainTest and used literal extension
strings (".json", ".js", ".yaml", ".yml") because MainTest sits in a
parent package and cannot reach Main's package-private
FILE_EXTENSION_* constants.

Per owner's request to use the constants directly:

- Reverts the additions to MainTest (helpers, 6 tests, imports of
  Hamcrest not/startsWith).
- Adds a focused new test class
  src/test/java/net/ladenthin/bitcoinaddressfinder/cli/MainConfigurationLoadingTest.java
  living in the same package as Main, so the @VisibleForTesting
  package-private constants Main.FILE_EXTENSION_JSON / _JS / _YAML /
  _YML are reachable directly.
- All six round-trip + syntax-probe assertions now reference the
  constants symbolically rather than the literal extension strings.
- MainTest's existing 14 tests remain untouched.

All tests pass: 6 in MainConfigurationLoadingTest (none skipped) +
14 in MainTest (6 skipped on OpenCL-gated cases as before).
